FalconSafetyApp2/requirements.txt
streamlit
ultralytics
opencv-python
Pillow
torch
gdown
